Move user creation to main
