+++ app/main.py
@@ -8,11 +8,12 @@ from fastapi import Depends, FastAPI, HTTPException, Request, status
 from fastapi.security import OAuth2PasswordRequestForm
 
 from app.database import get_mongo_client
-from app.models.user import Token
-from app.routers.auth import authenticate_user, create_access_token
+from app.models.user import CreateUser, Token
+from app.routers.auth import authenticate_user, create_access_token, hash_password
 from app.routers.auth import router as auth_router
 from app.routers.todo import router as todo_router
 from app.routers.user import router as user_router
+from app.utils.constants import FAILED_TO_CREATE_USER
 from app.utils.health import app_check_health, check_app_readiness
 
 app = FastAPI()
@@ -69,3 +70,17 @@ def login_for_access_token(
         data={"sub": user.username}, expires_delta=access_token_expires
     )
     return {"access_token": access_token, "token_type": "bearer"}
+
+
+@app.post("/user", response_model=bool)
+def create_user(username: str, email: str, password: str, request: Request):
+    hashed_password = hash_password(password)
+    user = CreateUser(username=username, email=email, hashed_password=hashed_password)
+    result = request.app.user.insert_one(user.dict())
+    if result.acknowledged:
+        return True
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail=FAILED_TO_CREATE_USER,
+        )
